Added component for account header values.
