+++ src/views/account/AccountHeaderChart.vue
@@ -1,67 +1,14 @@
 <template>
-  <div class="row bg-white">
-    <div class="col-12 col-md-6 px-1 py-3 px-md-3">
-      <div class="d-flex justify-content-between align-items-center border-bottom">
-        <span
-          class="com-font-s14-w700"
-          v-text="$t('labels.availables')"
-        />
-        <span
-          class="com-font-s14-w400"
-          v-text="availablesAmount"
-        />
-      </div>
-      <div class="d-flex justify-content-between align-items-center border-bottom">
-        <span
-          class="com-font-s14-w700"
-          v-text="$t('labels.delegated')"
-        />
-        <span
-          class="com-font-s14-w400"
-          v-text="delegationsAmount"
-        />
-      </div>
-      <div class="d-flex justify-content-between align-items-center border-bottom">
-        <span
-          class="com-font-s14-w700"
-          v-text="$t('labels.unbonded')"
-        />
-        <span
-          class="com-font-s14-w400"
-          v-text="unbondingDelegationsAmount"
-        />
-      </div>
-      <div class="d-flex justify-content-between align-items-center border-bottom">
-        <span
-          class="com-font-s14-w700"
-          v-text="$t('labels.rewards')"
-        />
-        <span
-          class="com-font-s14-w400"
-          v-text="rewardsAmount"
-        />
-      </div>
-      <div class="py-2 d-flex justify-content-between align-items-center com-font-s16-w700">
-        <span v-text="$t('labels.total')" />
-        <span v-text="totalsAmount" />
-      </div>
-    </div>
-    <div class="col-12 col-md-6 px-1 py-3 px-md-3">
-      <DoughnutChart
-        :chartdata="chartdata"
-        :options="options"
-        height="125"
-        width="125"
-      />
-    </div>
-  </div>
+  <DoughnutChart
+    :chartdata="chartdata"
+    :options="options"
+    height="120"
+    width="120"
+  />
 </template>
 
 <script>
 import DoughnutChart from "Components/common/DoughnutChart.vue";
-
-import { coinConverter } from "Utils";
-import { mapGetters } from "vuex";
 
 export default {
   name: "AccountHeaderChart",
@@ -70,21 +17,11 @@ export default {
     DoughnutChart
   },
   props: {
-    delegations: {
-      type: Array,
-      required: true,
-      note: "Delegations list"
-    },
-    rewards: {
-      type: Array,
-      required: true,
-      note: "Rewards list"
-    },
-    unbondings: {
-      type: Array,
-      required: true,
-      note: "Unbonding delegations list"
-    }
+    availables: { type: Number, default: 0, note: "Total token availables" },
+    delegated: { type: Number, default: 0, note: "Total token delegated" },
+    rewards: { type: Number, default: 0, note: "Total token rewards" },
+    unbonded: { type: Number, default: 0, note: "Total token unbonded" },
+    totals: { type: Number, default: 0, note: "Total tokens" }
   },
   data() {
     return {
@@ -94,11 +31,6 @@ export default {
         this.$t("labels.unbonded"),
         this.$t("labels.rewards")
       ],
-      totalAvailables: 0,
-      totalDelegated: 0,
-      totalUnbonded: 0,
-      totalRewards: 0,
-      chartdata: null,
       options: {
         responsive: true,
         maintainAspectRatio: false,
@@ -121,112 +53,14 @@ export default {
     };
   },
   computed: {
-    ...mapGetters("account", {
-      balances: "balances"
-    }),
-    denom() {
-      return this.balances && this.balances.length > 0
-        ? this.balances[0].denom
-        : "";
-    },
-    availablesAmount() {
-      let amount = {
-        denom: "",
-        amount: 0
-      };
-      if (this.balances && this.balances.length > 0) {
-        amount = coinConverter(this.balances[0]);
-        this.totalAvailables = parseFloat(this.balances[0].amount);
-      }
-      let formatAmount = this.$n(amount.amount, {
-        style: "decimal",
-        minimumFractionDigits: 6,
-        maximumFractionDigits: 6
-      });
-      return `${formatAmount} ${amount.denom}`;
-    },
-    delegationsAmount() {
-      this.delegations.forEach(element => {
-        this.totalDelegated += parseFloat(element.shares);
-      });
-      let amount = coinConverter({
-        denom: this.denom,
-        amount: this.totalDelegated
-      });
-      let formatAmount = this.$n(amount.amount, {
-        style: "decimal",
-        minimumFractionDigits: 6,
-        maximumFractionDigits: 6
-      });
-      return `${formatAmount} ${amount.denom}`;
-    },
-    rewardsAmount() {
-      this.rewards.forEach(element => {
-        this.totalRewards += parseFloat(element.amount);
-      });
-      let amount = coinConverter({
-        denom: this.denom,
-        amount: this.totalRewards
-      });
-      let formatAmount = this.$n(amount.amount, {
-        style: "decimal",
-        minimumFractionDigits: 6,
-        maximumFractionDigits: 6
-      });
-      return `${formatAmount} ${amount.denom}`;
-    },
-    unbondingDelegationsAmount() {
-      this.unbondings.forEach(element => {
-        element.entries.forEach(entry => {
-          this.totalUnbonded += parseFloat(entry.balance);
-        });
-      });
-      let amount = coinConverter({
-        denom: this.denom,
-        amount: this.totalUnbonded
-      });
-      let formatAmount = this.$n(amount.amount, {
-        style: "decimal",
-        minimumFractionDigits: 6,
-        maximumFractionDigits: 6
-      });
-      return `${formatAmount} ${amount.denom}`;
-    },
-    totalsAmount() {
-      let amount = coinConverter({
-        denom: this.denom,
-        amount: this.totals
-      });
-      let formatAmount = this.$n(amount.amount, {
-        style: "decimal",
-        minimumFractionDigits: 6,
-        maximumFractionDigits: 6
-      });
-      return `${formatAmount} ${amount.denom}`;
-    },
-    totals() {
-      return (
-        this.totalAvailables +
-        this.totalDelegated +
-        this.totalRewards +
-        this.totalUnbonded
-      );
-    }
-  },
-  watch: {
-    totals(value) {
-      if (value > 0) this.setChartdata();
-    }
-  },
-  methods: {
-    setChartdata() {
+    chartdata() {
       let data = [
-        this.getPercent(this.totalAvailables, this.totals),
-        this.getPercent(this.totalDelegated, this.totals),
-        this.getPercent(this.totalUnbonded, this.totals),
-        this.getPercent(this.totalRewards, this.totals)
+        this.getPercent(this.availables, this.totals),
+        this.getPercent(this.delegated, this.totals),
+        this.getPercent(this.unbonded, this.totals),
+        this.getPercent(this.rewards, this.totals)
       ];
-      this.chartdata = {
+      return {
         labels: this.labels,
         datasets: [
           {
@@ -235,9 +69,11 @@ export default {
           }
         ]
       };
-    },
+    }
+  },
+  methods: {
     getPercent(value, total) {
-      return ((value / total) * 100).toFixed(2);
+      return total > 0 ? ((value / total) * 100).toFixed(2) : 0;
     }
   }
 };
